Split out and generalize SPSCQueue, fix some alignments.
Disables dumping queues on error for now.
